+++ src/bittytax/conv/mergers/snowtrace.py
@@ -5,10 +5,8 @@ from typing import TYPE_CHECKING, Dict, List
 
 from ...bt_types import FileId
 from ..datamerge import DataMerge, ParserRequired
-from ..out_record import TransactionOutRecord
-from ..parsers.etherscan import etherscan_nfts, etherscan_tokens
-from ..parsers.snowtrace import WALLET, WORKSHEET_NAME, avax_int, avax_txns
-from .etherscan import INTERNAL_TXNS, NFTS, TOKENS, TXNS, _do_merge_etherscan
+from ..parsers.snowtrace import avax_tokens, avax_txns
+from .etherscan import TOKENS, TXNS, _do_merge_etherscan
 
 STAKE_ADDRESSES: List[str] = []
 
@@ -18,38 +16,14 @@ if TYPE_CHECKING:
 
 def merge_snowtrace(data_files: Dict[FileId, "DataFile"]) -> bool:
     # Do same merge as Etherscan
-    merge = _do_merge_etherscan(data_files, STAKE_ADDRESSES)
-
-    if merge:
-        # Change Etherscan wallet/worksheet name to SnowTrace
-        if TOKENS in data_files:
-            data_files[TOKENS].parser.worksheet_name = WORKSHEET_NAME
-            for data_row in data_files[TOKENS].data_rows:
-                if data_row.t_record:
-                    address = data_row.t_record.wallet[-abs(TransactionOutRecord.WALLET_ADDR_LEN) :]
-                    data_row.t_record.wallet = f"{WALLET}-{address}"
-
-                data_row.worksheet_name = WORKSHEET_NAME
-
-        if NFTS in data_files:
-            data_files[NFTS].parser.worksheet_name = WORKSHEET_NAME
-            for data_row in data_files[NFTS].data_rows:
-                if data_row.t_record:
-                    address = data_row.t_record.wallet[-abs(TransactionOutRecord.WALLET_ADDR_LEN) :]
-                    data_row.t_record.wallet = f"{WALLET}-{address}"
-
-                data_row.worksheet_name = WORKSHEET_NAME
-
-    return merge
+    return _do_merge_etherscan(data_files, STAKE_ADDRESSES)
 
 
 DataMerge(
-    "SnowTrace fees & multi-token transactions",
+    "Snowtrace fees & multi-token transactions",
     {
         TXNS: {"req": ParserRequired.MANDATORY, "obj": avax_txns},
-        TOKENS: {"req": ParserRequired.OPTIONAL, "obj": etherscan_tokens},
-        NFTS: {"req": ParserRequired.OPTIONAL, "obj": etherscan_nfts},
-        INTERNAL_TXNS: {"req": ParserRequired.OPTIONAL, "obj": avax_int},
+        TOKENS: {"req": ParserRequired.MANDATORY, "obj": avax_tokens},
     },
     merge_snowtrace,
 )